изменение структуры
