+++ src/Assets/Styles/App.module.scss
@@ -1,0 +1,4 @@
+.siderContent {
+    display: flex;
+    
+}
