+++ src/Assets/Styles/Content.module.scss
@@ -1,21 +1,5 @@
-.siteLayout {
-    padding: 0 50px;
-}
-
-.breadcrumb {
-    margin: 16px 0;
-}
-
 .container {
-    min-height: calc(100vh - 170px);
-}
-
-.sider {
-    width: 200px;      
-    padding: 23px 0;
-}
-
-.content {
-    padding: 24px;
-    min-height: 280px;
+    min-height: calc(100vh - 180px);
+    padding: 30px 60px;
+    width: 100%;
 }
