+++ src/Assets/Styles/Footer.module.scss
@@ -1,0 +1,3 @@
+.footer {
+    height: 50px;
+}
