--- src/Assets/Styles/MainButton.module.scss
@@ -1,9 +1,0 @@
-.wrapper{
-    text-align: center;
-}
-
-.button {
-    width: 200px;
-    text-align: center;
-    font-weight: 600;
-}
