+++ src/Assets/Styles/Sider.module.scss
@@ -1,0 +1,14 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    flex-wrap: nowrap;
+    justify-content: flex-start;
+    align-items: center;
+    align-content: center; 
+
+    padding: 30px 0 0px 30px;
+}
+
+.button {
+    width: 80px;
+}
